Create hook useMedications
